requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
environs==9.5.0
Flask-Caching==2.1.0
redis==4.5.1
requests==2.28.2
SQLAlchemy==2.0.6
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
alembic==1.13.1
boto3==1.26.100
Flask-Cors==3.0.10
gunicorn==21.2.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
pymongo==4.6.0
docker==6.1.3
autopep8==2.0.4
isort==5.13.2
SQLAlchemy-Utils==0.41.1
filetype==1.2.0
chardet==5.2.0
faiss-cpu==1.10.0
langchain-community==0.3.17
langchain-openai==0.3.5
scikit-learn==1.6.1
numpy==1.26.4